requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
openai==1.3.0
python-dotenv==1.0.0
faiss-cpu==1.7.4
psycopg2-binary==2.9.9
pypdf==3.17.1
docx2txt==0.8
flask==3.0.0
flask-socketio==5.3.6
python-socketio==5.10.0
python-engineio==4.8.0
werkzeug==3.0.1
sqlalchemy>=2.0.0
geoalchemy2>=0.14.0
python-docx>=1.0.0
PyPDF2>=3.0.0
openpyxl>=3.1.2
tiktoken>=0.5.1 